Refactor AuthContext and NewPasswordForm to include error handling during sign-out; add new dashboard components for revenue reporting and statistics visualization.
